=stop being so damn LAZYgit add *

+++ chapter-1/project-6/main.c
@@ -10,13 +10,13 @@ int power(int x, int y)
 
 int main(void)
 {
-	int x, result;
+	float x, result;
 	
 	printf("Enter a value for the variable x : ");
-	scanf("%d", &x);
+	scanf("%f", &x);
 
 	result = ( ( ( ( 3 * x + 2 ) * x - 5 ) * x - 1 ) * x + 7 ) * x - 6;
-	printf("The result is %d\n", result); 
+	printf("The result is %.2f\n", result); 
 
 	return 0;
 }
